% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{valid quick draw category}
\format{An object of class \code{character} of length 345.}
\source{
\link{https://raw.githubusercontent.com/googlecreativelab/quickdraw-dataset/master/categories.txt}
}
\usage{
categories
}
\description{
a vector of valid quick draw category
}
\keyword{datasets}
