% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_categories.R
\name{qd_categories}
\alias{qd_categories}
\title{valid quick draw category}
\usage{
qd_categories()
}
\value{
vector of valid quick draw category
}
\description{
valid quick draw category
}
\details{
The source is: \href{https://raw.githubusercontent.com/googlecreativelab/quickdraw-dataset/master/categories.txt}{raw_data}
}
