% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_download.r
\name{qd_download}
\alias{qd_download}
\title{qd_download download QuickDraw ndjson file}
\usage{
qd_download(category)
}
\arguments{
\item{category}{a valid QuickDraw category}
}
\value{
The file path to the ndjson
}
\description{
qd_download download QuickDraw ndjson file
}
\details{
Copy the ndjson file from Google Cloud to the cache directory
}
\examples{
qd_download("apple")
}
