% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_raster.R
\name{qd_raster}
\alias{qd_raster}
\title{qd_raster}
\usage{
qd_raster(data)
}
\arguments{
\item{data}{a tibble object resulted from qd_tidy()}
}
\value{
a tibble object with x and y coordinate of the raster matrix
}
\description{
qd_raster
}
\examples{
qd_read("apple") \%>\% tidy(1) \%>\% qd_raster()
}
