% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_read.R
\name{qd_read}
\alias{qd_read}
\title{qd_read}
\usage{
qd_read(category)
}
\arguments{
\item{category}{a valid QuickDraw category}
}
\value{
A tibble object contains 6 columns: word, countrycode, timestamp, recognized, key_id and drawing.
}
\description{
qd_read
}
\details{
Read in the QuickDraw ndjson file downloaded by qd_download()
}
\examples{
qd_read("apple")
}
