% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_sample.R
\name{qd_sample}
\alias{qd_sample}
\title{qd_sample}
\usage{
qd_sample(categories, size)
}
\arguments{
\item{categories}{a vector of valid QuickDraw category}

\item{size}{a vector of number of rows to take from each category}
}
\value{
A tibble object that binds the sample data from all catgories.
}
\description{
qd_sample
}
\details{
qd_sample() downloads QuickDraw data from multiple categories and take a sample from each category
}
\examples{
qd_sample(c("octopus", "star"), 5)
}
