% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_tidy.R
\name{qd_tidy}
\alias{qd_tidy}
\title{qd_tidy}
\usage{
qd_tidy(object, item = 1:nrow(object))
}
\arguments{
\item{object}{A tibble object produced by qd_read()}

\item{id}{The row number of drawing to tidy}
}
\value{
A tibble object with the x and y coordinate of each stroke
}
\description{
qd_tidy
}
\examples{
apple <- qd_read("apple")
qd_tidy(apple,1)
}
