% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickdraw-package.R
\docType{package}
\name{quickdraw-package}
\alias{quickdraw}
\alias{quickdraw-package}
\title{quickdraw: An API for pulling QuickDraw data from Google}
\description{
The package provides an API to download, read and cache QuickDraw data from (here)[https://github.com/googlecreativelab/quickdraw-dataset]. It also provides functionality for plotting and animating the individual drawing based on the coordinates stored in the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/quickdraw}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/quickdraw/issues}
}

}
\author{
\strong{Maintainer}: Sherry Zhang \email{huizezhangsh@gmail.com}

Authors:
\itemize{
  \item Stuart Lee [contributor]
}

}
\keyword{internal}
