% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_download.r
\name{qd_download}
\alias{qd_download}
\alias{qd_download_bitmap}
\title{qd_download download QuickDraw files}
\usage{
qd_download(category)

qd_download_bitmap(category)
}
\arguments{
\item{category}{a valid QuickDraw category}
}
\value{
A character vector containing
}
\description{
qd_download download QuickDraw files
}
\details{
\itemize{
\item \code{qd_download()}: retrieves and caches drawings in ndjson file
\item \code{qd_download_bitmap()}: retrieves and caches drawings in numpy bitmap format.
}
}
\examples{
qd_download("apple")
}
