% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_draw.R
\name{qd_draw}
\alias{qd_draw}
\title{qd_draw}
\usage{
qd_draw(object)
}
\arguments{
\item{object}{A tibble object produced by qd_tidy()}
}
\value{
An animation of the quick draw
}
\description{
qd_draw
}
\examples{
apple <- qd_read("apple") \%>\% qd_tidy(1)
qd_draw(apple)
}
